requirements.txt
aioredis==2.0.1
aioredis_fastapi==1.0.1
aiosmtplib==3.0.2
alembic==1.15.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
async-timeout==5.0.1
asyncio-redis==0.16.0
asyncpg==0.30.0
attrs==23.2.0
bcrypt==4.3.0
blinker==1.9.0
build==1.2.2
CacheControl==0.14.0
capstone==5.0.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.8
cloudinary==1.44.0
colorama==0.4.6
comtypes==1.4.7
contourpy==1.3.1
crashtest==0.4.1
cycler==0.12.1
distlib==0.3.8
Django==5.1.2
django-autoslug==1.9.9
django-crispy-forms==2.3
django-jazzmin==3.0.1
django-taggit==6.1.0
django-tinymce==4.1.0
djangorestframework==3.15.2
dnspython==2.7.0
dulwich==0.21.7
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.105.0
fastapi-cli==0.0.7
fastapi-mail==1.4.2
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.0.3
fonttools==4.55.8
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.6
keyring==24.3.1
kiwisolver==1.4.8
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.0
numpy==2.2.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pefile==2023.2.7
pexpect==4.9.0
pillow==11.0.0
pkginfo==1.11.1
platformdirs==4.3.6
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.4.8
Pybag==2.2.12
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pygame==2.6.1
pygbag==0.9.2
Pygments==2.19.1
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pyjokes==0.8.3
PyJWT==2.10.1
pyparsing==3.2.1
pypiwin32==223
pyproject_hooks==1.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pyttsx3==2.97
pytz==2025.2
pywin32==306
pywin32-ctypes==0.2.3
PyYAML==6.0.2
rapidfuzz==3.9.7
redis==6.0.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.3
rsa==4.9.1
selenium==4.27.1
setuptools==75.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.40
sqlparse==0.5.1
starlette==0.27.0
timeout-decorator==0.5.0
tomlkit==0.13.2
trio==0.27.0
trio-websocket==0.11.1
trove-classifiers==2024.9.12
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.2
urllib3==2.4.0
uv==0.4.26
uvicorn==0.34.2
virtualenv==20.26.5
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.4
wsproto==1.2.0